- Rename TEST_ONLT to UNIT_TEST - Clean up
