+++ backup/cortex/CMakeLists.txt
@@ -18,6 +18,25 @@ MACRO( GET_SUBDIR result curdir)
 ENDMACRO()
 
 
+if ($ENV{METOPE_EMULATOR} EQUAL 1)
+    if( UNIT_TEST AND UNIT_TEST EQUAL 1)
+        message(STATUS "[CMakeLists] File operation: $ENV{PRJ_TOP}/backup/main_unittest.cc => $ENV{PRJ_TOP}/main.cc")
+        file(COPY_FILE "$ENV{PRJ_TOP}/backup/main_unittest.cc" "$ENV{PRJ_TOP}/main.cc")
+    else()
+        message(STATUS "[CMakeLists] File operation: $ENV{PRJ_TOP}/backup/main_emulator.cc => $ENV{PRJ_TOP}/main.cc")
+        file(COPY_FILE "$ENV{PRJ_TOP}/backup/main_emulator.cc" "$ENV{PRJ_TOP}/main.cc")
+    endif()
+else()
+    if ($ENV{METOPE_CHIP} STREQUAL "NATIVE")
+        message(STATUS "[CMakeLists] File operation: $ENV{PRJ_TOP}/backup/main_native.cc => $ENV{PRJ_TOP}/main.cc")
+        file(COPY_FILE "$ENV{PRJ_TOP}/backup/main_native.cc" "$ENV{PRJ_TOP}/main.cc")
+    else()
+        message(STATUS "[CMakeLists] File operation: $ENV{PRJ_TOP}/backup/main_program.cc => $ENV{PRJ_TOP}/main.cc")
+        file(COPY_FILE "$ENV{PRJ_TOP}/backup/main_program.cc" "$ENV{PRJ_TOP}/main.cc")
+    endif()
+endif()
+
+
 
 #########################################################################################################
 # Project Path & Source Files Definition
@@ -115,7 +134,6 @@ if($ENV{METOPE_EMULATOR} EQUAL 1)
 else()
     list(APPEND LINK_FLAG "-specs=nosys.specs")
 endif()
-                                    
 
 
 if((NOT IMU) OR (IMU STREQUAL "QMI8658A") )
@@ -168,13 +186,13 @@ endif()
 
 #########################################################################################################
 # Customized Project Macros
-# @param TEST_ONLY      - Test Bench Program was enabled. Escape the main program.
+# @param UNIT_TEST      - Test Bench Program was enabled. Escape the main program.
 #   @param INCLUDE_TB_BSP - Enable Human Interaction Test Bench (ie. Tests that need manual intervine)
 #   @param INCLUDE_TB_CMN - Enable Common Test Bench (ie. Basic utility functional tests)
 #   @param INCLUDE_TB_OS  - Enable OS Test Bench (ie. RTOS related Test Bench)
 #########################################################################################################
-if( TEST_ONLY AND TEST_ONLY EQUAL 1)
-    list(APPEND DEF_LIST "-DTEST_ONLY=1")
+if( UNIT_TEST AND UNIT_TEST EQUAL 1)
+    list(APPEND DEF_LIST "-DUNIT_TEST=1")
     if(INCLUDE_TB_BSP AND INCLUDE_TB_BSP EQUAL 1)
         list(APPEND DEF_LIST "-DINCLUDE_TB_BSP=1")
     endif()
