+++ backup/main_emulator.cc
@@ -1,8 +1,8 @@
 /**
  ******************************************************************************
- * @file    main.cc
+ * @file    main_emulator.cc
  * @author  RandleH
- * @brief   Unit Testing
+ * @brief   Emulator Customization Program
  ******************************************************************************
  * @attention
  *
@@ -21,7 +21,6 @@
 /* ************************************************************************** */
 /*                                  Includes                                  */
 /* ************************************************************************** */
-
 #include <stdio.h>
 #include "init.h"
 #include "trace.h"
