+++ setup.env
@@ -82,14 +82,11 @@ if [[ "$1" = native ]]; then
   export METOPE_CHIP="NATIVE"
   export METOPE_EMULATOR=0
   cp ${PRJ_TOP}/backup/native/CMakeLists.txt ${PRJ_TOP}/
-  cp ${PRJ_TOP}/backup/main_native.cc ${PRJ_TOP}/main.cc
 else
   if [[ "$2" = emulator ]]; then
     export METOPE_EMULATOR=1
-    cp ${PRJ_TOP}/backup/main_emulator.cc ${PRJ_TOP}/main.cc
   else
     export METOPE_EMULATOR=0
-    cp ${PRJ_TOP}/backup/main_program.cc ${PRJ_TOP}/main.cc
   fi
 
   if [[ "$1" = STM32F405RGT6 ]]; then
