+++ tool/rebuild.sh
@@ -24,17 +24,18 @@ cd ${PRJ_TOP}/build
 
 # cmake -DCMAKE_BUILD_TYPE=Debug ..
 
-# cmake -DCMAKE_BUILD_TYPE=Debug -DTEST_ONLY=1 ..
+# cmake -DCMAKE_BUILD_TYPE=Debug -DUNIT_TEST=1 ..
 
 # cmake -DCMAKE_BUILD_TYPE=Debug ..
 
-# cmake -DCMAKE_BUILD_TYPE=Debug -DTEST_ONLY=1 -DINCLUDE_TB_OS=1  ..
+# cmake -DCMAKE_BUILD_TYPE=Debug -DUNIT_TEST=1 -DINCLUDE_TB_OS=1  ..
 
-cmake -DCMAKE_BUILD_TYPE=Debug -DTEST_ONLY=1 -DINCLUDE_TB_CMN=1 -DQMI8658A=1 ..
+# cmake -DCMAKE_BUILD_TYPE=Debug -DUNIT_TEST=1 -DINCLUDE_TB_CMN=1 -DQMI8658A=1 ..
+cmake -DCMAKE_BUILD_TYPE=Debug -DUNIT_TEST=1 -DINCLUDE_TB_CMN=1 -DQMI8658A=1 ..
 
-# cmake -DCMAKE_BUILD_TYPE=Debug -DTEST_ONLY=1 -DINCLUDE_TB_BSP=1 -DQMI8658A=1 ..
+# cmake -DCMAKE_BUILD_TYPE=Debug -DUNIT_TEST=1 -DINCLUDE_TB_BSP=1 -DQMI8658A=1 ..
 
-# cmake -DCMAKE_BUILD_TYPE=Debug -DTEST_ONLY=1 -DINCLUDE_TB_OS=1 ..
+# cmake -DCMAKE_BUILD_TYPE=Debug -DUNIT_TEST=1 -DINCLUDE_TB_OS=1 ..
 
 ################################################################
 # Execution Zone
